Rename vm -> compiler
