+++ funsor/compiler.py
